Make sdk works in browser and node environments
